Percents, svg height and hidden blobs
Percents shouldn't be formatted, so we can use them in math.

The <svg> shouldn't be higher than necessary.

Separate the concept of things that are hidden because other things
are zoomed (.zoom-unwanted), and the concept of things that are
hidden because they are to thin (.zoom-too-thin). This way, we
can search among .zoom-too-thin and not .zoom-unwanted.